+++ lib/TaskProcessing/EvaSummaryProvider.php
@@ -25,7 +25,7 @@ class EvaSummaryProvider implements ISynchronousProvider {
 	}
 
 	public function getName(): string {
-		return $this->l->t('Eva (local)');
+		return $this->l->t('Eva · Local');
 	}
 
 	public function getTaskTypeId(): string {